implemement basic game play
game #take_turn method implemented which lead to implementing a #to_s
method for the player class. Also implemented basic #play method
that asks for moves with switching players each time. It also checks for
a game over condition and announces the result, both of which are
still to be implemented